🚚 api: Move certification-course controller to session-management context
